+++ lib/cocoapods/installer/xcode/pods_project_generator.rb
@@ -4,7 +4,7 @@ module Pod
       # The {PodsProjectGenerator} handles generation of the 'Pods/Pods.xcodeproj'
       #
       class PodsProjectGenerator
-        require 'cocoapods/installer/pods_project_integrator/pod_target_integrator'
+        require 'cocoapods/installer/xcode/pods_project_generator/pod_target_integrator'
         require 'cocoapods/installer/xcode/pods_project_generator/target_installer'
         require 'cocoapods/installer/xcode/pods_project_generator/pod_target_installer'
         require 'cocoapods/installer/xcode/pods_project_generator/file_references_installer'
@@ -63,7 +63,7 @@ module Pod
           prepare
           install_file_references
           install_libraries
-          integrate_test_targets
+          integrate_targets
           set_target_dependencies
         end
 
@@ -179,12 +179,12 @@ module Pod
           end
         end
 
-        def integrate_test_targets
+        def integrate_targets
           pod_targets_with_test_targets = pod_targets.reject { |pt| pt.test_native_targets.empty? }
           unless pod_targets_with_test_targets.empty?
-            UI.message '- Integrating test targets' do
+            UI.message '- Integrating targets' do
               pod_targets_with_test_targets.each do |pod_target|
-                Pod::Installer::PodTargetIntegrator.new(pod_target).integrate!
+                PodTargetIntegrator.new(pod_target).integrate!
               end
             end
           end